Remove unused code and improve error handling.
The commit removes an unused import and a commented-out line from BottomController.java. Additionally, code related to image duplication search has been removed from TableUtils.java as it was unused and verbose. To improve error handling, an additional message has been added in TableUtils.java and the corresponding string entry in lang_en.properties file for instances when the ConflictTable cannot be loaded.