Refactor code and adjust UI settings
The commit mainly makes two types of changes. On one hand, it simplifies a mathematical calculation in the 'StageControl.java' file to make the code cleaner and easier to read. On the other hand, it modifies certain UI settings in the 'Main.fxml' file, making the 'tables_hbox' always grow in 'VBox' and modifying the way rotation of image views are handled in the 'TableController.java' file. It also removes an unused variable 'hide_im'.